Added conversion script for pre-release state files.
The master branch is frozen until first release (0.5.0) at the end of
March. This script will convert state files made before this release
to the format used for that release.

No tables were modified, but now uuid, containertype, and container name
are reflected in the filesystem, not in the state file. These entries
are no longer used if they are present.