path2container.py now recognizes full paths.
Also, added fixes needed to ensure containertype correctly determined by
looking at the basename of the full path to the file, not the full path.
In other words, 'Sim' must be in the state file's basename, not the full
path to it, to qualify it as a Sim. This was a very insidious bug.