Refactoring GroupFile given variable member names.
Given that Container names are now mutable, it makes zero sense to store
these in the Group's state file. The column has been removed from the
Members table. Need to look closely to ensure this causes no problems
for backwards compatibility, since we will need an updater script if
this breaks anything.

The refactoring also makes member data access more efficient, allowing
for multiple path checks when looking for members. The changes to the
GroupFile toward this functionality have yet to be reflected in the
Members aggregator, which will implement the rest.