Fixes #20. Added band-aid to catch corner case HDFStore doesn't.
When storing Pandas objects that don't have labels (for DataFrames,
columns) with non-numerical values, the pdDataFile object stores the
object and indexes its labels, so querying can be done on those
elements. However if DataFrame columns have only numerical names, then
querying can't be done on these elements and indexes won't be built.

This commit catches a corner case in which a 1-D DataFrame with a
numerical column name shouldn't be stored with indexes built. Hopefully
this band-aid can be removed in the future if the HDFStore object itself
can catch this instead.