Combined pandas fixtures for tests.
Ideally we would have liked to have been able to use separate fixtures
for each pandas structure we wanted to test, but there is no clear way
of iterating through these and running the same tests on each. A way
might exist, but it's opaque to me at the moment. Looping through
a dictionrary of data structures for each test works for now.