Hacking on MemberData.
The MemberData aggregator is intended to give Members access to data
elements of members as aggregates. First such method is retrieve, which
only works for the moment with dataframes. Will need to expand its
functionality to work with anything we support.

Still, it automatically builds a MultiIndex dataframe from every
dataframe matching the given handle in each member. Neat!