Hammering away at Bundle.
In order to do a proper job of extending Member's capabilities, we want
to do it in a way that transfers directly to Bundle. Also, we want Group
indexing to start returning Bundles, which requires them to be up to
snuff. Working on making this so.

Bundle and Members share practically the same API, but they inherently
have different underlying backends (Bundle is memory-only; Members are
linked to a GroupFile on disk). To make it possible to write one good
set of methods for both, keeping things simpler and consistent between
them, I've made a _BundleBackend class that serves as the backend to
Bundle, with the same interface as the member-specific elements of
GroupFile. I've also started building a mixin called _CollectionBase
that should contain the common elements for each of Bundle and Members.