Updated and standardized sizes of string columns.
Unfortunately there is no easy way of storing variable-length strings in
a pytables Table, so the solution for now is to set some proactively
long lengths for things like paths and tags, categories, etc. We should
be able to change this in the future without breaking existing
Containers, and it may be worthwhile implementing a mechanism to expand
the length of a string column when a string is too large to fit in an
existing one.