Bundle backend now stores member data in structured array.
We want to use the same interface for both Members and Bundle, and since
we take advantage of the fact that the data stored in the GroupFile
member table can be quickly returned as a structured array, it made
sense to structure the in-memory backend for Bundle as a structured
array. This should give some performance gains over what we were using
(a list of dicts), and makes for a more consistent low-level API anyhow.