Addressed PEP8 violations in a number of files.
We now have py.test use the pep8 plugin to flag areas where code doesn't
follow a consistent style, and this has pointed to a host of violations.
Most of these are either::

    * line too long
    * continuation at inconsistent indent level
    * whitespace on blank line

I addressed these in a number of files, but there are stil more to go.
It's mostly just tedious work, but should be done to help ensure future
code that is merged into master is readable and has consistent style.