Finished up work on Foxhound usage for Groups and Bundles.
This was a major undertaking. Now, when Groups go looking for their
members they feed the uuids and paths they have stored for these members
to a Foxhound object. On fetch, the Foxhound looks in the paths given to
it for all uuids it's tasked with finding, and for those members it
can't find it walks the filesystem tree starting from the Group's basedir,
then upward to each succeeding parent directory. It does this until all
members are found, it exhaustively searches the filesystem, or until it
times out (default 100 seconds).

Also, slicing a Group's members returns a Bundle, which has the same
interface as Group.Members but stores its state information in memory.
Both Bundle and Members run from the same interface code, but use
different backends (obviously), so anything one can do, the other can
do as well.

Bundles also find their members with a Foxhound, and this appears to
work as expected. The main difference is that the upward trajectory for
the search begins from the current working directory, since the Bundle
has no disk representation.