More tests for Sims and Groups, plus taking advantage of inheritance.
Pytest makes it fairly easy to use inheritance for test classes, and
this makes sense to use for elements common to Containers, Sims, and
Groups (namely, all of Container's stuff). Although in principle testing
only that Container works in all these areas is sufficient, it's
important that Group and Sim actually manage to work there, too. Now
any tests that are built for Container are also built for Sims and
Groups.