Removed logger info for adding universe from SimFile.
We still don't have a standard for what level logging should happen at
(interface vs. backend), or for which loglevels. However, because
add_universe is commonly used just to update paths, even when a Universe
is built with Sim.universes.__getitem__, we don't care to see a log
message spewing out this information.

Perhaps as a proposed measure: loglevel info should be reserved for
interface methods, since they are closest to the context of what is
being done by the user. Backends should probably stick to loglevel
warning and above only. Need to add this as a discussion issue.