Fixes #28
The numpy array data getter used slice notation to grab the full array
from the h5py file instance, but this was hacky enough to faile for the
case of scalar arrays (length 0). Instead, replacing the slicing with
``.value`` works for all array types.

Added a test for a scalar array.