Fixes #30
We should avoid using append mode ('ab') for storing pickle objects, as this
doesn't work for dumping with pickle. Changed instead to using write
mode ('wb'), since there are no instances in which the concept of
appending to a pickle is something we want to dive into.

Thanks to @cing (Chris Ing) for the fix.