Fixes Issue #10
Containers and their aggregators should be able to work with different
backends, which may very well not be PyTables files. To make this
possible, we need to catch PyTables-specific exceptions at the File
level, and raise more generic exceptions to be caught at the
Container/aggregator level.

We now follow the convention that backend-level exceptions that are
specific to the implementation of that backend need to be caught and
re-raised as either built-in exceptions or MDS exceptions. Containers
and aggregators must then catch and re-raise the appropriate
interface-level exception with an appropriate interface-level message.

For the changes made in this commit, these re-raises look a bit silly
since the messages and exceptions are identical to those caught. We
leave them in, however, to indicate the clear separation between backend
and frontend, and that these exceptions/messages may differ between the
two levels.

Added seleciton tests.