Fixes #16
Universes now check both stored absolute and relative file paths when
reloading. This ensures that Containers that are moved will still be
able to load their Universes as before so long as either the relative or
absolute paths remain the same. This is especially useful for
trajectories that are stored within a Container's tree, for example.