Replaced selections table with pytables array.
Because we store a selection wholesale, and modifying it amounts to
destroying the table and writing a new one, we don't need these to
be pytables Table nodes. Instead, they can be simple arrays. This avoids
having to worry about setting a string length for all selections
(although this could have been handled with a method that returns a
 Table definition given an input string size), and it also results in
cleaner code.

It should also be more efficient, since storing large selections, which
is often the case for selections using AtomGroups (integer indices are
        stored), we avoid having to do a python loop to deposit each
row. Instead, the array is written in one go.