Reorganized tests.
Moved the various data structures we try to store and retrieve to their
own submodule so that they can be pulled in individually by other tests.
This was necessary so that we can use the same data structures when
retrieving aggregated datasets from collections without having to
reinvent them elsewhere.

Also split out the tests for Group.Members into a mixin called
CollectionTests. This mixin is found in test_bundle, and it corresponds
to the machinery found in _CollectionBase. It is written so that it can
be used to test both Group.Members and Bundle with the same code.

Working on MemberData tests now. This is tricky because we have to
practically encode the content of the ``MemberData.retrieve()`` method
to test that the data structures are built correctly.