The core classes of MDSynthesis are now datreant.
The developers of MDSynthesis (https://github.com/Becksteinlab/MDSynthesis/issues/38)
have decided to split out its core into a new package called "datreant".
This makes the unique way datreant handles the problem of heterogeneous
data storage and recall available to users in fields outside of
molecular dynamics work, hopefully spurring development in new and
wonderful ways. MDSynthesis will live on as a package built around
datreant with MDAnalysis-specific subclasses of datreant Treants.

One note of interest: to reduce blandness and avoid confusion with other
common uses of the word, all instances of "Container" have been replaced
with "Treant". For the uninitiated, a treant is a mythical walking,
talking tree, originating the land of Dungeons and Dragons. The idea is
that datreant's data model relies on directory trees, and leverages
their usefulness over database-only solutions.

If the use of 'Treant' and 'treant' in the documentation results in too
much confusion, we will consider other names.

The directory package of MDSynthesis has been reworked to accommodate
this change. For one thing, it's been flattened. If this turns out to be
confusing, we'll partition it up. One worry is that it removes the
visual separation between the backend (File objects) and the frontend
(Treants).