Locks appear to work now. :D
So, it appears (though I'm still not exactly sure this is what was
happening) that applying the file locks AFTER opening the state file
for reading or writing was problematic, probably because of caching
on the part of PyTables as soon as the file is opened. However,
one needs a file descriptor to apply a lock, and there is no way (as
far as I can tell) of of feeding a file descriptor for an HDF5 file
to a pytables function that would then return a tables.File object.

To get around this, I added a File.fd attribute to the File class
in datreant. This keeps an active file descriptor for the state
file on which advisory locks (through lockf) can be applied. Once
obtained, the file is opened using pytables (creating another file
descriptor in the same process). After any reading/writing is done,
the file is closed and the lock is removed from the file descriptor.

At the moment the file descriptor is kept alive for the life of the
TreantFile object. It might be wise to make it exist only when needed
for an imminent read/write?