Hacking on pdData to ensure locks work with multiprocessing.
There appears to be a race condition somewhere when trying to append
to the same stored dataframe from multiple processes. It appears the
condition rears its head more often when the dataframes are small. In
other words, there might be some slippage when more than on process
tries to acquire a lock at the same time.