Append locks for pandas objects appears solved.
I think the problem with appending to pandas objects with
multiprocessing was that on init of pdDataFile an HDFStore instance
with the append flag was opened and immediately closed. This might have
created some subtle stale caches with the Groups present in the file
(which is what it looks like the errors were indicating; looked like
 spew from the hdflib, and the files couldn't be opened by vitables
 or pytables).

To avoid this operation on init, we create an empty, hidden proxy file
in the same directory and apply our locks to file descriptors opened
on it, instead of on the actual data file. This sidesteps the issue
of which process will create the data file first, since the one that
does will have already gotten a lock on a file no one cares about.