datreant collections are now extendable.
In order for datreant collections like Group and Bundle to know how to
handle other types of Treants in the filesystem, the Treants
and their backend classes can be added to the dictionaries
datreant._treants and datreant._treantfiles. We may make these
unhidden and give them their own namespace in the future to make
them more accessible.