Can now add and remove members of a collection with glob syntax.
Bundles and Groups are intended to be pretty dynamic, in that adding and
removing members shouldn't be painful. To make this easy, members can be
added directly from the filesystem with file glob patterns. They can
also be removed by name or by glob patterns applied against the names
of all members.

Fixes #4.