Multiple backends now supported; improved registration mechanism.
It's now possible to use other backends besides the built-in pytables
ones for treants, although none exist yet. These can be added to the
corresponding treant's _backends dictionary, which is used to
grab the appropriate backend based on the encountered file extension. We
have yet to expose a kwarg to choose what backend to use on new treant
creation, however.

The registration mechanism for new treants is also simplified using
the top-lvel `register` function.

Also, adjusted the version and schema update mechanism. Now versions of
files shouldn't update until we add in the machinery to update their
schema, too. This does not exist yet.