Fixed issue with Foxhound finding proxy files first on Travis.
When state files go missing, the Foxhound is employed to go looking for
them. However, it was misidentifying proxy files as state files on
Travis, which wasn't happening on my development machine. This might
have to do with what scandir finds first on the given system, which
depends on the system. The fix should keep this from happening by
excluding hidden files from candidacy.