Fixes #14.
Removed 'location' keyword from Treant init. Also removed init method
from Group and therefore the ability to add members through the
constructor. This comes with the advantage of not having to edit Treant
children whenever the init method changes.

It's now no longer necessary to use a separate keyword to put a Treant
in another directory. What used to be it's 'name' is now just the path
to the Treant, which will always include its name (the basedir of the
statefile). This simplifies creating and reloading Treants, since the
same syntax works for both. See the documentation for the constructor
for the details.