Hammering away on new package structure.
Trying to figure out how to refactor data storage scheme. At the moment
it doesn't expand well to other data structures we may want to handle in
the future; it's a bit hard-coded for what it can do now. I think
conceptually something like blaze's `Data` interface is something that
would work well in the long term, and we might give some way of getting
back a `Data` object for a stored dataset in the future, but blaze isn't
really usable yet in the ways we would like. The hdfstore interface, for
example, doesn't give an easy way to just read in the data. We don't
want to be dependent on blaze development to keep datreant useful.

Let's try and make do with what we can, without worrying too much about
alternate backends.