Still hacking away on reorganization.
This is all a bit complicated. Whatever the final form, we may pull
pieces of it into the main branch. For now, this is rather low priority,
and it complicates the package. Considering how we don't have clean
solutions to extending the data storage mechanism to other
structures/formats, I think we can only approach a half solution. The
only thing gained at the moment would be separation of backends for
state files.

HOWEVER, another advantage which we need nowish is the ability to define
schema updates on a per-Treant and per-format basis. If we try to do
this in the current scheme, especially with the experimental yaml
backend, it will be a bit messy.