We now also have a JSON backend!
Created a `TreantFileSerial` base class for serialization formats like
YAML and JSON, in which the whole file most be deserialized and
reserialized on each change. Defining YAML and JSON TreantFiles then
only requires definition of the `_deserialize` and `_serialize` methods.

The value in these is that the HDF5 format is performant and scales
to large amounts of metadata, but requires the HDF5 libs and isn't human
readable. The JSON format concedes speed and scaling for human readability,
while the YAML format concedes more speed for more human readability. I
think this ability to choose is quite cool.

Todo yet: make a serial backend for Groups.