Fixed! Inconsistency in how GroupFile*.get_members() returned.
Now `get_members` for all GroupFile variants give a dictionary of lists
instead of a structured array. This sacrifices a bit of performance but
allows us to avoid dealing with fixed-length strings for formats like
JSON and YAML. It turns out this method is used only once in the
codebase by higher-level objects anyhow.