Removed datreant.data.
This is an experiment to see how well having separate submodules that
dynamically add limbs to class definitions works. Added methods for
adding limbs to Treants and Collections, and we use these to dynamically
add tags and categories, for example, to Treants and its subclasses,
as well as members to Groups.