Limbs now have a backend and logger property.
If a Treant swaps out its statefile object for a new one, and if its
limbs have references to the old state file, this will cause proplems.
However, it's useful for limbs and treants to be able to refer to the
statefile with the same name (self._backend), since this makes objects
like Foxhound easier to generalize. We now use a property for this in
the Limb base class to achieve this effect.