Now use metaclassing to automatically register Treants, Limbs, and AggLimbs
It's no longer necessary to explicitly register objects for datreant to
be aware of them; sublcassing automatically does this.

Treants and collections now have an `attach` method that takes the name
of the limb to attach (name must be in registry, which happens
automatically on import). Attaching a limb to a collection attaches it
to all its members.