Now we also have TreeLimbs.
Limbs like Tags and Categories depend on a state file, so they must be
attached to a Treant. However, other limbs, such as datreant.data's Data
limb, don't depend on a state file, and so can be attached to a plain
Tree. This allows subselection of directory structures and data inside
as suggested in #23.