Trees no longer make directories automatically.
Trees and Leaves now behave like raw paths, meaning making them does not create
them in the filesystem. Added a `makedirs` method to each that creates
all directories needed to describe the path. This results in Tree.exists
giving True in the case of a Tree, but for Leaf this of course does not
make the last component of the path since this is supposed to be a File.