Working on several things at once.
Finally shoring up some technical debt by adding tests for Trees and
Leaves, for which the former also apply to Treants. Not finished with
these yet. Also:

1. Adding in AggTags.__getitem__, which allows for quick selection of
members from the Bundle of Treants having one or more tags. Will try to
do and/or selection with lists/tuples, respectively.

2. Bundles will finally behave like sets with the usual operators: union
(| or +), intersection (&), difference (-), symmetric difference (^),
and subset (<, <=) and superset (>, >=). We do want to add in the
appropriate methods to make duck-typing wrt sets fully work for these.

3. AggCategories is nowhere near done. It's more complicated than
AggTags, but the basic design is taking cues from what works for it.