TreeLimbs added to child classes of Tree now make it to Tree.
For a limb like datreant.data.Data, we need to be able to attach it to
Tree subclasses and be available to Tree without having to attach it to
the Tree class, because this would mean that all uses of Treant have
this automatically, which may be undesirable. Instead, now the calling
Tree-derived class instance will make Tree instances upon __getitem__
that have the TreeLimbs it does.