Group.__repr__ no longer has to find members.
It was a problem if one loaded a Group with Sims as members without
having imported mdsynthesis. Although members won't work right at the
moment (need to default to Treant on failure to find 'Sim') if the
treanttype of a member is not understood, this also ensures Groups
can be used interactively even if their members can't fully.