We now have AggTreeLimbs.
Since Trees are capable of having attachable limbs (see datreant.data
for an example) that don't require a state file (and therefore a
Treant), it makes sense that Views should get aggregation limbs that
correspond to these. AggTreeLimbs are limbs that are attachable to both
Views and Bundles by design. By contrast, AggLimbs are for Bundles only.