I like this better. Now View does behave a bit as an abstract Tree.
We want View to act somewhat like a Tree that has its members as
children. So, when we call View.trees, we expect to get the members of
the View that are Trees, just as when we call Tree.trees we get the
children of the Tree that are Trees. Also, indexing by name in a View is
roughly similar to indexing by name in a Tree (though not exactly; for
Views only names of actual are allowed).

Where we diverge a bit is with the behavior of View.children.
Tree.children will given all the children of that Tree. View.children
will give all the children of the Trees within the View. There isn't a
way around this because in order for Tree.children.children to work as
one expects, View.children must not just give back itself. So...the
conceptual idea of a View as an abstract Tree works mostly, but not
fully. Partial duck-typing, in a way.