Now have both filter and glob methods in View.
View is a bit confused at the moment. There are two competing views
(heh) for how View should behave: 1) as a collection like Bundle whose API
mirrors that of the objects it contains, giving convenient aggregations
of their individual functionality, or 2) as an abstract Tree that has
its members as children. It's confusing trying to make it behave roughly
as both, and there are places where we can't have it both ways.

I'm leaning toward 1 right now. Preparing the way for that on a separate
branch and we'll see what sticks.