Lots of doc updates; moved API references to own section.
The autodoc API references can get pretty overwhelming, so moved to
their own main section of the docs. We link to them in each part of the
user documentation as it already exists.

Removed Bundle.discover. It really just looked like a point of
confusion, and is unnecessary since discover is now a top-level function
in the library.